+++ libshvhistory/rpchistory.c
@@ -1,7 +1,6 @@
 #include <stdlib.h>
 #include <shv/rpchistory.h>
 
-#include "getlog_param_keys.gperf.h"
 #include "shvc_config.h"
 
 const struct rpcdir rpchistory_fetch = {
@@ -48,8 +47,11 @@ const struct rpcdir rpchistory_lastsync = {
 bool rpchistory_getlog_response_pack_begin(
 	cp_pack_t pack, struct rpchistory_record_head *head, int ref) {
 	cp_pack_imap_begin(pack);
-	cp_pack_int(pack, RPCHISTORY_GETLOG_KEY_TIMESTAMP);
-	cp_pack_datetime(pack, head->datetime);
+
+	if (head->datetime.msecs != INT64_MAX) {
+		cp_pack_int(pack, RPCHISTORY_GETLOG_KEY_TIMESTAMP);
+		cp_pack_datetime(pack, head->datetime);
+	}
 
 	if (ref >= 0) {
 		cp_pack_int(pack, RPCHISTORY_GETLOG_KEY_REF);
@@ -93,11 +95,11 @@ bool rpchistory_getlog_response_pack_end(cp_pack_t pack) {
 static bool unpack_value(cp_unpack_t unpack, struct cpitem *item,
 	struct obstack *obstack, struct rpchistory_getlog_request *res, int key) {
 	switch (key) {
-		case K_SINCE:
+		case RPCHISTORY_GETLOG_REQ_KEY_SINCE:
 			return cp_unpack_datetime(unpack, item, res->since);
-		case K_UNTIL:
+		case RPCHISTORY_GETLOG_REQ_KEY_UNTIL:
 			return cp_unpack_datetime(unpack, item, res->until);
-		case K_COUNT:
+		case RPCHISTORY_GETLOG_REQ_KEY_COUNT:
 			switch (cp_unpack_type(unpack, item)) {
 				case CPITEM_INT:
 					return cpitem_extract_int(item, res->count);
@@ -107,22 +109,19 @@ static bool unpack_value(cp_unpack_t unpack, struct cpitem *item,
 				default:
 					return false;
 			}
-		case K_SNAPSHOT:
-			return cp_unpack_bool(unpack, item, res->snapshot);
-		case K_RI:
+		case RPCHISTORY_GETLOG_REQ_KEY_RI:
 			res->ri = cp_unpack_strdupo(unpack, item, obstack);
 			if (res->ri == NULL)
 				res->ri = "**:*";
 			return true;
-		default: /* skip values for unknown or unsupported keys */
-			cp_unpack_skip(unpack, item);
-			return true;
+		default:
+			return false;
 	}
 }
 
 struct rpchistory_getlog_request *rpchistory_getlog_request_unpack(
 	cp_unpack_t unpack, struct cpitem *item, struct obstack *obstack) {
-	if (cp_unpack_type(unpack, item) != CPITEM_MAP)
+	if (cp_unpack_type(unpack, item) != CPITEM_IMAP)
 		return NULL;
 
 	struct rpchistory_getlog_request *res = obstack_alloc(obstack, sizeof(*res));
@@ -132,29 +131,21 @@ struct rpchistory_getlog_request *rpchistory_getlog_request_unpack(
 		.since = {.msecs = (ts.tv_sec * 1000) + (ts.tv_nsec / 1000000)},
 		.until = {.msecs = (ts.tv_sec * 1000) + (ts.tv_nsec / 1000000)},
 		.count = -1,
-		.snapshot = false,
 		.ri = "**:*",
 	};
 
-	for_cp_unpack_map(unpack, item, key, 10) {
-		const struct gperf_rpchistory_getlog_param_key_match *match =
-			gperf_rpchistory_getlog_param_key(key, strlen(key));
-		if (!match)
-			cp_unpack_skip(unpack, item);
-		else if (!unpack_value(unpack, item, obstack, res, match->key)) {
+	for_cp_unpack_imap(unpack, item, key) {
+		if (!unpack_value(unpack, item, obstack, res, key)) {
 			obstack_free(obstack, res);
 			return NULL;
 		}
 	}
 
-	if (res->count == UINT_MAX && res->snapshot)
+	if (res->since.msecs == res->until.msecs)
 		res->count = 0;
 
 	if (res->count > SHVC_GETLOG_LIMIT)
 		res->count = SHVC_GETLOG_LIMIT;
-
-	if (res->since.msecs >= res->until.msecs)
-		res->snapshot = false;
 
 	return res;
 }